+++ apps/prairielearn/python/prairielearn/__init__.py
@@ -1,1 +1,4 @@
+from prairielearn.attribute_utils import *  # noqa: F403
+from prairielearn.conversion_utils import *  # noqa: F403
 from prairielearn.core import *  # noqa: F403
+from prairielearn.validation_utils import *  # noqa: F403
