+++ apps/prairielearn/python/prairielearn/validation_utils.py
@@ -1,0 +1,111 @@
+"""Utility functions for validating student data."""
+
+from typing import Any
+
+import numpy as np
+import numpy.typing as npt
+from numpy.typing import ArrayLike
+
+
+def is_int_json_serializable(n: int) -> bool:
+    return -((2**53) - 1) <= n <= 2**53 - 1
+
+
+# This is a deprecated alias that will be removed in the future -- use the lowercase version instead.
+def is_correct_ndarray2D_dd(*args: Any, **kwargs: Any) -> bool:  # noqa: N802
+    return is_correct_ndarray2d_dd(*args, **kwargs)
+
+
+# This is a deprecated alias that will be removed in the future -- use the lowercase version instead.
+def is_correct_ndarray2D_sf(*args: Any, **kwargs: Any) -> bool:  # noqa: N802
+    return is_correct_ndarray2d_sf(*args, **kwargs)
+
+
+# This is a deprecated alias that will be removed in the future -- use the lowercase version instead.
+def is_correct_ndarray2D_ra(*args: Any, **kwargs: Any) -> bool:  # noqa: N802
+    return is_correct_ndarray2d_ra(*args, **kwargs)
+
+
+def is_correct_ndarray2d_dd(
+    a_sub: npt.NDArray[Any], a_tru: npt.NDArray[Any], digits: int = 2
+) -> bool:
+    # Check if each element is correct
+    m = a_sub.shape[0]
+    n = a_sub.shape[1]
+    for i in range(m):
+        for j in range(n):
+            if not is_correct_scalar_dd(a_sub[i, j], a_tru[i, j], digits):
+                return False
+
+    # All elements were close
+    return True
+
+
+def is_correct_ndarray2d_sf(
+    a_sub: npt.NDArray[Any], a_tru: npt.NDArray[Any], digits: int = 2
+) -> bool:
+    # Check if each element is correct
+    m = a_sub.shape[0]
+    n = a_sub.shape[1]
+    for i in range(m):
+        for j in range(n):
+            if not is_correct_scalar_sf(a_sub[i, j], a_tru[i, j], digits):
+                return False
+
+    # All elements were close
+    return True
+
+
+def is_correct_ndarray2d_ra(
+    a_sub: npt.NDArray[Any],
+    a_tru: npt.NDArray[Any],
+    rtol: float = 1e-5,
+    atol: float = 1e-8,
+) -> bool:
+    # Check if each element is correct
+    return np.allclose(a_sub, a_tru, rtol, atol)
+
+
+def is_correct_scalar_ra(
+    a_sub: ArrayLike, a_tru: ArrayLike, rtol: float = 1e-5, atol: float = 1e-8
+) -> bool:
+    """Compare a_sub and a_tru using relative tolerance rtol and absolute tolerance atol."""
+    return bool(np.allclose(a_sub, a_tru, rtol, atol))
+
+
+def is_correct_scalar_dd(a_sub: ArrayLike, a_tru: ArrayLike, digits: int = 2) -> bool:
+    """Compare a_sub and a_tru using digits many digits after the decimal place."""
+    # If answers are complex, check real and imaginary parts separately
+    if np.iscomplexobj(a_sub) or np.iscomplexobj(a_tru):
+        real_comp = is_correct_scalar_dd(a_sub.real, a_tru.real, digits=digits)  # type: ignore
+        imag_comp = is_correct_scalar_dd(a_sub.imag, a_tru.imag, digits=digits)  # type: ignore
+        return real_comp and imag_comp
+
+    # Get bounds on submitted answer
+    eps = 0.51 * (10**-digits)
+    lower_bound = a_tru - eps
+    upper_bound = a_tru + eps
+
+    # Check if submitted answer is in bounds
+    return bool((a_sub > lower_bound) & (a_sub < upper_bound))
+
+
+def is_correct_scalar_sf(a_sub: ArrayLike, a_tru: ArrayLike, digits: int = 2) -> bool:
+    """Compare a_sub and a_tru using digits many significant figures."""
+    # If answers are complex, check real and imaginary parts separately
+    if np.iscomplexobj(a_sub) or np.iscomplexobj(a_tru):
+        real_comp = is_correct_scalar_sf(a_sub.real, a_tru.real, digits=digits)  # type: ignore
+        imag_comp = is_correct_scalar_sf(a_sub.imag, a_tru.imag, digits=digits)  # type: ignore
+        return real_comp and imag_comp
+
+    # Get bounds on submitted answer
+    if a_tru == 0:
+        n = digits - 1
+    else:
+        n = -int(np.floor(np.log10(np.abs(a_tru)))) + (digits - 1)
+    eps = 0.51 * (10**-n)
+    lower_bound = a_tru - eps
+    upper_bound = a_tru + eps
+
+    # Check if submitted answer is in bounds
+    return bool((a_sub > lower_bound) & (a_sub < upper_bound))
